pyproject.toml
[project]
name = "ai-agents-from-scratch"
version = "0.1.0"
description = "Learn to build AI agents from scratch - no frameworks, just fundamentals"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openai>=1.12.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "tiktoken>=0.5.0",
    "chromadb>=0.4.0",
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "sse-starlette>=1.8.0",
    "gunicorn>=21.2.0",
    "python-multipart>=0.0.6",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "ipykernel>=6.0.0",
]

[dependency-groups]
dev = [
    "pytest-asyncio>=1.2.0",
    "pytest-cov>=7.0.0",
]

[tool.pytest.ini_options]
markers = [
    "integration: Integration tests that make real API calls (deselect with '-m \"not integration\"')"
]
addopts = "-v"
